[#26] refactor: service 단의 인터페이스들을 한 파일에 통합
