+++ services/auth_service.go
@@ -8,12 +8,6 @@ import (
 	"golang.org/x/crypto/bcrypt"
 )
 
-type AuthService interface {
-	SignUp(user *models.User) error
-	Login(email, password string) (string, error) // TODO: JWT 토큰 리턴 수정
-}
-
-// authService 구조체
 type authService struct {
 	userRepository repositories.UserRepository
 }
@@ -22,7 +16,6 @@ func NewAuthService(userRepo repositories.UserRepository) AuthService {
 	return &authService{userRepository: userRepo}
 }
 
-// SignUp - 새로운 사용자 등록
 func (s *authService) SignUp(user *models.User) error {
 	// TODO: 이메일 중복시 클라이언트단에 알림
 	_, err := s.userRepository.GetUserByEmail(user.Email)
@@ -37,11 +30,9 @@ func (s *authService) SignUp(user *models.User) error {
 	}
 	user.Password = string(hashedPassword)
 
-	// 사용자 저장
 	return s.userRepository.CreateUser(user)
 }
 
-// Login - 이메일과 비밀번호로 로그인 처리
 func (s *authService) Login(email, password string) (string, error) {
 	user, err := s.userRepository.GetUserByEmail(email)
 	if err != nil {
